backend/requirements.txt
# ATLAS v2 Backend Dependencies

# Core framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
python-dotenv>=1.0.0

# Geospatial
osmnx>=2.0.0
networkx>=3.4
geopandas>=1.0.0
shapely>=2.0.6
scikit-learn>=1.5.0

# Google Earth Engine
earthengine-api>=1.7.0

# Data processing
numpy>=2.1.0

# HTTP client
httpx>=0.27.0

# Database
sqlalchemy>=2.0.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0

# PDF reports
reportlab>=4.0.0
